requirements.txt
feedparser>=6.0.11
Jinja2>=3.1.3
openai>=1.30.0
PyYAML>=6.0.1
